XMMATRIX、無事死亡！！！！！！！

+++ Util.h
@@ -3,7 +3,7 @@
 namespace Util
 {
 	//ウィンドウサイズ
-	const int window_width = 1280;
-	const int window_height = 720;
+	const float window_width = 1280;
+	const float window_height = 720;
 };
