+++ WinAPI.cpp
@@ -45,7 +45,7 @@ void WinAPI::SetWindowClass()
 	RegisterClassEx(&w);
 
 	//ウィンドウサイズ { X座標 Y座標 横幅 縦幅 }
-	wrc = { 0,0,Util::window_width,Util::window_height };
+	wrc = { 0,0,(long)Util::window_width,(long)Util::window_height };
 
 	//自動でサイズを補正する
 	AdjustWindowRect(&wrc, WS_OVERLAPPEDWINDOW, false);
